rag_final/RAG_task/parser/requirements.txt
PyPDF2
python-docx
langchain==0.2.16
langchain-community==0.0.25
langchain-core==0.2.38
langchain-openai
faiss-cpu
python-dotenv
sentence-transformers